+++ src/components/notice/PartyNotice.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { Link, useNavigate } from 'react-router-dom';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
-import { noticeState } from '../atoms';
+import { noticeState } from '../../atoms';
 
 // 이미지 import
 import { ReactComponent as LeftBack } from '../assets/chating/leftBack.svg';
@@ -11,7 +11,7 @@ import { ReactComponent as Alarm } from '../assets/notice/alarm.svg';
 import { ReactComponent as Close } from '../assets/notice/close.svg';
 import { ReactComponent as NullAlram } from '../assets/notice/nullalram.svg';
 import { ReactComponent as Email } from '../assets/notice/email.svg';
-import { deleteAPI, getAPI, postAPI } from '../api/api';
+import { deleteAPI, getAPI, postAPI } from '../../api/api';
 
 export const PartyNotice = () => {
   const navigate = useNavigate();
